+++ include/RemoteDetonator.hpp
@@ -5,7 +5,7 @@
 
 class RemoteDetonator : public Drop {
 	public:
-		RemoteDetonator(std::pair<int, int> xy);
+		RemoteDetonator(std::pair<float, float> xy);
 		~RemoteDetonator();
 		RemoteDetonator(RemoteDetonator const & src);
 		RemoteDetonator	&operator=(RemoteDetonator const &src);
