+++ include/UnbreakableBox.hpp
@@ -8,8 +8,8 @@ class UnbreakableBox : public Box {
 public:
 	UnbreakableBox();
 	UnbreakableBox(Drop *drop);
-	UnbreakableBox(std::pair<int, int> xy);
-	UnbreakableBox(std::pair<int, int> xy, Drop *drop);
+	UnbreakableBox(std::pair<float, float> xy);
+	UnbreakableBox(std::pair<float, float> xy, Drop *drop);
 	~UnbreakableBox();
 	UnbreakableBox(UnbreakableBox const &src);
 	UnbreakableBox		&operator=(UnbreakableBox const &src);
